\name{CEGxls-package}
\alias{CEGxls-package}
\docType{package}
\title{
A collection of utility functions for manipulating xls files.
}
\description{
A collection of utility functions for manipulating xls files. 
}
\details{
\tabular{ll}{
Package: \tab CEGxls\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2009-02-14\cr
License: \tab CEG\cr
}

}
\author{
Adrian Dragulescu.
}
\references{
}
\keyword{ package }
\examples{
}
